rebuild package-lock.json with current versions (#2)
Purpose
Avoid vulnerabilities in transitive dependencies.

Approach
Reinstall without a package-lock.json file in order to rebuild it from
scratch, picking up newly-released versions, resolving the most
egregious warnings.

TODOS and Open Questions
Two vulnerabilities remain:
* nth-check <2.0.1 (high) :: css-select <=3.1.0 :: svgo 1.0.0 - 1.3.2 ::
  @svgr/plugin-svgo <=5.5.0 :: @svgr/webpack 4.0.0 - 5.5.0 ::
  react-scripts >=2.1.4.
* postcss <8.4.31 (moderate) :: resolve-url-loader@"^4.0.0" ::
  react-scripts@5.0.1

We could try adding an overrides entry for @svgr/webpack to v6 (the
breaking change was dropping support for webpack v4, but this repo is
already on v5, so that may not affect us at all). IIRC the postcss v8
changes are significant so I'm not sure we have a good option there.